+++ net-spider/ChangeLog.md
@@ -4,6 +4,10 @@
 
 * Add Weaver module.
 
+### Unify module
+
+* Add `toLinkSamples` function.
+
 ## 0.4.1.0  -- 2020-01-26
 
 * Add `GraphML.Attribute` module. `AttributeValue` and related types