+++ net-spider/ChangeLog.md
@@ -1,6 +1,6 @@
 # Revision history for net-spider
 
-## 0.4.2.0  -- ???
+## 0.4.2.0  -- 2020-03-21
 
 * Add Weaver module.